requirements.txt
absl-py==2.1.0
aiofiles==23.2.1
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyascii==0.3.2
anyio==3.7.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-lru==2.0.2
async-timeout==4.0.3
attrs==23.1.0
audiocraft @ git+https://github.com/facebookresearch/audiocraft@69fea8b290ad1b4b40d28f92d1dfc0ab01dbab85
audioread==3.0.1
av==11.0.0
Babel==2.12.1
backcall==0.2.0
bangla==0.0.2
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.7.0
blis==0.7.11
bnnumerizer==0.0.2
bnunicodenormalizer==0.1.6
boto3==1.34.40
botocore==1.34.40
catalogue==2.0.10
certifi==2019.11.28
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.1.0
click==8.1.7
cloudpathlib==0.16.0
cloudpickle==3.0.0
colorama==0.4.6
colorlog==6.8.2
comm==0.1.3
confection==0.1.4
contourpy==1.2.0
coqpit==0.0.17
cycler==0.12.1
cymem==2.0.8
Cython==3.0.8
dateparser==1.1.8
dbus-python==1.2.16
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
demucs==4.0.1
dnspython==2.5.0
docopt==0.6.2
dora_search==0.1.12
einops==0.7.0
email-validator==2.1.0.post1
encodec==0.1.1
exceptiongroup==1.1.1
executing==1.2.0
fastapi==0.109.2
fastjsonschema==2.17.1
ffmpy==0.3.2
filelock==3.13.1
flashy==0.0.2
Flask==3.0.2
fonttools==4.48.1
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.2.0
funcy==2.0
g2pkk==0.1.2
gradio==4.18.0
gradio_client==0.10.0
grpcio==1.60.1
gruut==2.2.3
gruut-ipa==0.13.0
gruut-lang-de==2.0.0
gruut-lang-en==2.0.0
gruut-lang-es==2.0.0
gruut-lang-fr==2.0.2
h11==0.14.0
hangul-romanize==0.1.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
huggingface-hub==0.20.3
hydra-colorlog==1.2.0
hydra-core==1.3.2
idna==2.8
importlib-resources==6.1.1
inflect==7.0.0
ipykernel==6.23.3
ipython==8.14.0
isoduration==20.11.0
itsdangerous==2.1.2
jamo==0.4.1
jedi==0.18.2
jieba==0.42.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
json5==0.9.14
jsonlines==1.2.0
jsonpointer==2.4
jsonschema==4.17.3
julius==0.2.7
jupyter-events==0.6.3
jupyter-lsp==2.2.0
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyter_server==2.7.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.2
jupyterlab-pygments==0.2.2
jupyterlab_server==2.23.0
kiwisolver==1.4.5
lameenc==1.7.0
langcodes==3.3.0
lazy_loader==0.3
librosa==0.10.0
lightning-utilities==0.10.1
llvmlite==0.42.0
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.1
more-itertools==10.2.0
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.5
murmurhash==1.0.10
mutagen==1.47.0
nbclient==0.8.0
nbconvert==7.6.0
nbformat==5.9.0
nest-asyncio==1.5.6
networkx==2.8.8
nltk==3.8.1
notebook_shim==0.2.3
num2words==0.5.13
numba==0.59.0
numpy==1.22.0
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
openai-whisper==20231117
openunmix==1.2.1
orjson==3.9.13
overrides==7.3.1
packaging==23.1
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pillow==10.2.0
platformdirs==3.8.0
pooch==1.8.0
preshed==3.0.9
prometheus-client==0.17.0
prompt-toolkit==3.0.38
protobuf==4.25.2
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pydantic==2.6.1
pydantic-extra-types==2.5.0
pydantic-settings==2.1.0
pydantic_core==2.16.2
pydub==0.25.1
Pygments==2.15.1
PyGObject==3.36.0
pynndescent==0.5.11
pyparsing==3.1.1
pypinyin==0.50.0
pyrsistent==0.19.3
pysbd==0.3.4
python-apt==2.0.1+ubuntu0.20.4.1
python-crfsuite==0.9.10
python-dateutil==2.8.2
python-dotenv==1.0.1
python-json-logger==2.0.7
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0
pyzmq==25.1.0
regex==2023.12.25
requests==2.31.0
requests-unixsocket==0.2.0
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.0
ruff==0.2.1
s3transfer==0.10.0
safetensors==0.4.2
scikit-learn==1.4.0
scipy==1.11.4
semantic-version==2.10.0
Send2Trash==1.8.2
sentencepiece==0.1.99
shellingham==1.5.4
six==1.14.0
smart-open==6.4.0
sniffio==1.3.0
soundfile==0.12.1
soupsieve==2.4.1
soxr==0.3.7
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
stack-data==0.6.2
starlette==0.36.3
submitit==1.5.1
SudachiDict-core==20240109
SudachiPy==0.6.8
suno-bark @ git+https://github.com/suno-ai/bark.git@773624d26db84278a55aacae9a16d7b25fbccab8
sympy==1.12
tensorboard==2.16.0
tensorboard-data-server==0.7.2
terminado==0.17.1
tf_keras-nightly==2.16.0.dev2024021210
thinc==8.2.3
threadpoolctl==3.2.0
tiktoken==0.6.0
tinycss2==1.2.1
tokenizers==0.15.2
tomli==2.0.1
tomlkit==0.12.0
toolz==0.12.1
torch==2.1.0
torchaudio==2.1.0
torchmetrics==1.3.1
tornado==6.3.2
tqdm==4.66.2
trainer==0.0.36
traitlets==5.9.0
transformers==4.37.2
treetable==0.2.5
triton==2.2.0
TTS @ git+https://github.com/coqui-ai/TTS@dbf1a08a0d4e47fdad6172e433eeb34bc6b13b4e
typer==0.9.0
typing_extensions==4.9.0
tzdata==2024.1
tzlocal==5.2
ujson==5.9.0
umap-learn==0.5.5
Unidecode==1.3.8
uri-template==1.3.0
urllib3==1.25.8
uvicorn==0.27.1
uvloop==0.19.0
wasabi==1.1.2
watchfiles==0.21.0
wcwidth==0.2.6
weasel==0.3.4
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.1
websockets==11.0.3
Werkzeug==3.0.1
xformers==0.0.22.post7
yarl==1.9.4
